+++ src/main/java/org/huazhi/util/MenuTree.java
@@ -22,13 +22,13 @@ public class MenuTree extends TreeNode {
 	public MenuTree() {
 	}
 
-	public MenuTree(int id, String name, int parentId) {
+	public MenuTree(Long id, String name, Long parentId) {
 		this.id = id;
 		this.parentId = parentId;
 		this.name = name;
 	}
 
-	public MenuTree(int id, String name, MenuTree parent) {
+	public MenuTree(Long id, String name, MenuTree parent) {
 		this.id = id;
 		this.parentId = parent.getId();
 		this.name = name;
